+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Support `drag_source true` on `list` widget
 - Support `drop_target true` keyword alternative to `on_drop` for cases where it is implied what to drop (text in the case of label, entry, combo, list, and button)
+- Support `on('drag_start')` (instead of `on_drag_start`), `on('drag_motion')`, and `on('drop')` for consistency with Glimmer DSL for general Tk listener style (`on('event')`)
 
 ## 0.0.45